ejercicio3/requirements.txt
django==4.2.4
django-filter==23.2
mysqlclient==2.2.0
Pillow==10.0.0
djangorestframework==3.14.0
django-environ==0.10.0
django-rest-knox==4.2.0
requests==2.31.0
django-cors-headers==4.2.0